User Update API fixed
