+++ app/config/settings/base.py
@@ -80,6 +80,7 @@ AUTHENTICATION_BACKENDS = [
 
 CORS_ORIGIN_WHITELIST = (
     'himanmen.com',
+    'localhost:3000',
 )
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
