+++ app/members/models.py
@@ -1,26 +1,28 @@
 import os
-
 from django.conf import settings
+
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.db.models import Manager
 
+SIGNUP_TYPE_EMAIL = 'e'
+SIGNUP_TYPE_FACEBOOK = 'f'
+
+SIGNUP_TYPE_CHOICES = (
+    (SIGNUP_TYPE_FACEBOOK, 'facebook'),
+    (SIGNUP_TYPE_EMAIL, 'email'),
+)
+
 
 class User(AbstractUser):
-    SIGNUP_TYPE_EMAIL = 'e'
-    SIGNUP_TYPE_FACEBOOK = 'f'
-
-    SIGNUP_TYPE_CHOICES = (
-        (SIGNUP_TYPE_FACEBOOK, 'facebook'),
-        (SIGNUP_TYPE_EMAIL, 'email'),
-    )
+    file_path = os.path.join(settings.STATIC_DIR, 'img_profile_default.png')
 
     username = models.CharField(max_length=255, unique=True, blank=True, null=True)
     email = models.EmailField(max_length=255, unique=False, null=True)
 
-    img_profile = models.ImageField(upload_to='user', default='/static/iu.jpg')
+    img_profile = models.ImageField(upload_to='user', blank=True, default='/static/iu.jpg')
+    phone_num = models.CharField(max_length=20, blank=True)
     signup_type = models.CharField(max_length=1, choices=SIGNUP_TYPE_CHOICES, default=SIGNUP_TYPE_EMAIL)
-    phone_num = models.CharField(max_length=20, blank=True, null=True)
 
     created_date = models.DateField(auto_now_add=True)
     modified_date = models.DateField(auto_now=True)
