+++ app/members/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .auth import *
+from .facebook import *
+from .user import *
